% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{adaptFilePath}
\alias{adaptFilePath}
\title{Adapt file path. Change backslash to slash}
\usage{
adaptFilePath(path)
}
\arguments{
\item{path}{File location}
}
\value{
File location String using slash as separator
}
\description{
Fix paths so internal console recognize paths. Change backslash to backslash
}
\examples{
newPath <- adaptFilePath('\temp\path\to\file\here.tif')
newPath
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
