% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{ALL_PARTITION_METHOD}
\alias{ALL_PARTITION_METHOD}
\title{All supported partition methods}
\usage{
ALL_PARTITION_METHOD()
}
\value{
a vector of supported partition methods
}
\description{
All supported partition methods
}
\examples{
ALL_PARTITION_METHOD()
}
