\name{all_partition_methods}
\alias{all_partition_methods}
\title{
All supported partitioning methods
}
\description{
All supported partitioning methods
}
\usage{
all_partition_methods()
}
\details{
New partitioning methods can be registered by \code{\link{register_partition_methods}}.
}
\value{
A vector of supported partitioning methods.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
all_partition_methods()
}
