% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{ALL_TOP_VALUE_METHOD}
\alias{ALL_TOP_VALUE_METHOD}
\title{All supported top methods}
\usage{
ALL_TOP_VALUE_METHOD()
}
\value{
a vector of supported top methods.
}
\description{
All supported top methods
}
\examples{
ALL_TOP_VALUE_METHOD()
}
