\name{all_top_value_methods}
\alias{all_top_value_methods}
\title{
All supported top value methods
}
\description{
All supported top value methods
}
\usage{
all_top_value_methods()
}
\value{
A vector of supported top value methods.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
all_top_value_methods()
}
