\name{cal_consensus}
\alias{cal_consensus}
\title{
Calculate consensus clustering related statistics
}
\description{
Calculate consensus clustering related statistics
}
\usage{
cal_consensus(partition_list, k)
}
\arguments{

  \item{partition_list}{a list of partitions. The partitions can be vectors of categorical values or objects already generated by clue package (the hard partitions only).}
  \item{k}{number of classes that are expected.}

}
\value{
A list of following elements:

\describe{
  \item{class_ids}{consensus class ids.}
  \item{consensus_matrix}{consensus matrix.}
  \item{membership_mat}{probability to be in each group. }
  \item{membership_each}{original partition list that has been adjusted according to the consensus class ids.}
  \item{stat}{statistics for the consensus clustering}
}
}
\examples{
# There is no example
NULL
}
