% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{cdpop_mapdensity}
\alias{cdpop_mapdensity}
\title{Makes a population density map from CDPOP results interpolation}
\usage{
cdpop_mapdensity(
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/interpolate_popdensity.py"),
  grids,
  template,
  method = "average",
  bandwidths = "None",
  type = "count",
  crs = "None"
)
}
\arguments{
\item{py}{Python location}

\item{grids}{String. Vector with the CDPOP output files to be interpolated}

\item{template}{String. Raster template for interpolating the files}

\item{method}{String. Method for interpolation. Options are: 'isj', 'silvermans', 'scotts', 'average', 'cv', 'user'}

\item{type}{String.}

\item{crs}{String. User provided CRS as EPSG or ESRI string. Can also be 'None' in which case the CRS will be extracted from the template raster.}

\item{cdpopscript}{Python location}

\item{bandwiths}{String.}
}
\value{
List with three slots: a) file, with NA if no result given or a single file if function was successful, b) newFiles, string vector with resulting files, c)log, string with the message obtained from the console execution. 0 indicates success.
}
\description{
Takes CDPOP results and generates a raster interpolation
}
\examples{
cdpop_mapdensity( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
