% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{cdpop_mapstruct}
\alias{cdpop_mapstruct}
\title{Makes a population structure map from CDPOP results interpolation}
\usage{
cdpop_mapstruct(
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/interpolate_popstructure.py"),
  grids,
  template,
  method = "thin_plate_spline",
  neighbors,
  crs = "None",
  cml = TRUE,
  show.result = TRUE
)
}
\arguments{
\item{py}{Python executable location}

\item{pyscript}{Python script location}

\item{grids}{String, List of CDPOP grid.csv files containing population genetic structure}

\item{template}{String. Path/filename of a template raster used for interpolation}

\item{method}{String. Interpolation method. One of 'multiquadric', 'thin_plate_spline', 'linear', or 'idw'.}

\item{neighbors}{String. Number of neighbors to use for interpolation. Either a positive integer < the total number of occupied points or 'all'.}

\item{crs}{String. User provided CRS as EPSG or ESRI string. Can also be 'None' in which case the CRS will be extracted from the template raster.}
}
\value{
List with three slots: a) file, with NA if no result given or a single file if function was successful, b) newFiles, string vector with resulting files, c)log, string with the message obtained from the console execution. 0 indicates success.
}
\description{
Takes CDPOP results and generates a raster interpolation
}
\examples{
grid <- system.file(package = 'cola', 'sampledata/cdpop/grid0.csv')
template <- system.file(package = 'cola', 'sampledata/sampleHS.tif')

struMap <- cdpop_mapstruct(
  grids = grid, template = template,
  method = 'thin_plate_spline', neighbors = 'all', crs = 'None')

}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
