\name{cola}
\alias{cola}
\title{
A bottle of cola
}
\description{
A bottle of cola
}
\usage{
cola()
}
\details{
Simply serve you a bottle of cola.

The ASCII art is from \url{http://ascii.co.uk/art/coke} .
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
for(i in 1:10) cola()
}
