% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cola2.R
\docType{data}
\name{cola2_params}
\alias{cola2_params}
\title{Parameters for installation}
\format{
An object of class \code{list} of length 2.
}
\usage{
cola2_params
}
\description{
Defines the default cola installation parameters
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
\keyword{datasets}
