% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{cola_dss}
\alias{cola_dss}
\title{Launch \emph{COLA} decision support system (DSS) dashboard}
\usage{
cola_dss(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Run this app on a new window? Default `TRUE`}
}
\description{
This function runs the \emph{cola} dashboard
}
\note{
Please see the official website (\url{https://github.com/connectingLandscapes/cola/})
}
\examples{
library(cola)
if(interactive()) {
cola_dss()
}
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
