% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_params.R, R/setup.R
\docType{data}
\name{cola_params}
\alias{cola_params}
\title{Parameters for installation}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 5.
}
\usage{
cola_params

cola_params
}
\value{
The temperature in degrees Celsius
}
\description{
Default parameters for cola conda environment installation

Defines the default cola installation parameters
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
\keyword{datasets}
