\name{cola_report-ConsensusPartition-method}
\alias{cola_report,ConsensusPartition-method}
\title{
Make report for the ConsensusPartition object
}
\description{
Make report for the ConsensusPartition object
}
\usage{
\S4method{cola_report}{ConsensusPartition}(object, output_dir)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}
  \item{output_dir}{the path for the output html file}

}
\details{
A html report which contains all plots
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
