\name{cola_report-dispatch}
\alias{cola_report}
\title{
Method dispatch page for cola_report
}
\description{
Method dispatch page for \code{cola_report}.
}
\section{Dispatch}{
\code{cola_report} can be dispatched on following classes:

\itemize{
\item \code{\link{cola_report,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{cola_report,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{cola_report,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
