\name{collect_classes-dispatch}
\alias{collect_classes}
\title{
Method dispatch page for collect_classes
}
\description{
Method dispatch page for \code{collect_classes}.
}
\section{Dispatch}{
\code{collect_classes} can be dispatched on following classes:

\itemize{
\item \code{\link{collect_classes,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{collect_classes,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{collect_classes,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
