% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_classes.consensus_partition}
\alias{collect_classes.consensus_partition}
\title{Collect classes from consensus_partition object}
\usage{
\method{collect_classes}{consensus_partition}(x, show_legend = TRUE, ...)
}
\arguments{
\item{x}{a \code{consensus_partition} object}

\item{show_legend}{whether show legend.}

\item{...}{other arguments.}
}
\description{
Collect classes from consensus_partition object
}
