% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_classes.consensus_partition_all_methods}
\alias{collect_classes.consensus_partition_all_methods}
\title{Collect classes from consensus_partition_all_methods object}
\usage{
\method{collect_classes}{consensus_partition_all_methods}(x, k,
  top_method = x$top_method, partition_method = x$partition_method, ...)
}
\arguments{
\item{x}{a \code{consensus_partition_all_methods} object returned by \code{\link[=run_all]{run_all()}}.}

\item{k}{number of partitions}

\item{top_method}{a vector of top methods}

\item{partition_method}{a vector of partition methods}

\item{...}{other arguments.}
}
\description{
Collect classes from consensus_partition_all_methods object
}
