% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_classes}
\alias{collect_classes}
\title{General method for collect_classes}
\usage{
collect_classes(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{other arguments}
}
\description{
General method for collect_classes
}
