% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_classes.run_all}
\alias{collect_classes.run_all}
\title{Collect classes from run_all object}
\usage{
\method{collect_classes}{run_all}(x, k, top_method = x$top_method,
  partition_method = x$partition_method, ...)
}
\arguments{
\item{x}{a \code{run_all} object returned by \code{\link[=run_all]{run_all()}}.}

\item{k}{number of partitions}

\item{top_method}{a vector of top methods}

\item{partition_method}{a vector of partition methods}

\item{...}{other arguments.}
}
\description{
Collect classes from run_all object
}
