\name{collect_plots-dispatch}
\alias{collect_plots}
\title{
Method dispatch page for collect_plots
}
\description{
Method dispatch page for \code{collect_plots}.
}
\section{Dispatch}{
\code{collect_plots} can be dispatched on following classes:

\itemize{
\item \code{\link{collect_plots,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{collect_plots,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
