% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_plots.consensus_partition}
\alias{collect_plots.consensus_partition}
\title{Collect plots from consensus_partition object}
\usage{
\method{collect_plots}{consensus_partition}(x, ...)
}
\arguments{
\item{x}{a \code{consensus_partition} object}

\item{...}{other arguments}
}
\description{
Collect plots from consensus_partition object
}
