% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_plots}
\alias{collect_plots}
\title{General method for collect_plots}
\usage{
collect_plots(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{other arguments}
}
\description{
General method for collect_plots
}
