% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_plots.run_all}
\alias{collect_plots.run_all}
\title{Collect plots from run_all object}
\usage{
\method{collect_plots}{run_all}(x, k = 2, fun = consensus_heatmap,
  top_method = x$top_method, partition_method = x$partition_method, ...)
}
\arguments{
\item{x}{a \code{run_all} object from \code{run_all()}.}

\item{k}{number of partitions.}

\item{fun}{function used to generate plots. Valid functions are \code{\link[=consensus_heatmap]{consensus_heatmap()}},
\code{\link[=plot_ecdf]{plot_ecdf()}}, \code{\link[=membership_heatmap]{membership_heatmap()}} and \code{\link[=get_signatures]{get_signatures()}}.}

\item{top_method}{a vector of top methods.}

\item{partition_method}{a vector of partition methods.}

\item{...}{other arguments passed to corresponding \code{fun}.}
}
\description{
Collect plots from run_all object
}
