% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_stat}
\alias{collect_stat}
\title{Cophenetic correlation coefficient in all methods}
\usage{
collect_stat(res_list, top_method = res_list$top_method,
  partition_method = res_list$partition_method)
}
\arguments{
\item{res_list}{a \code{consensus_partition_all_methods} object}

\item{top_method}{a vector of top methods}

\item{partition_method}{a vector of partition methods}
}
\description{
Cophenetic correlation coefficient in all methods
}
\details{
A data frame with following columns:

-\code{cophcor}: the cophenetic correlation coefficient
-\code{mean_silhouette}: mean silhouette
-\code{concordance_to_known}: correlation to known factors if it is provided.
}
