\name{collect_stats-ConsensusPartition-method}
\alias{collect_stats,ConsensusPartition-method}
\title{
Draw and compare statistics for a single method
}
\description{
Draw and compare statistics for a single method
}
\usage{
\S4method{collect_stats}{ConsensusPartition}(object, ...)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object.}
  \item{...}{other arguments.}

}
\details{
It is identical to \code{\link{select_partition_number,ConsensusPartition-method}}.
}
\examples{
# There is no example
NULL
}
