\name{collect_stats-dispatch}
\alias{collect_stats}
\title{
Method dispatch page for collect_stats
}
\description{
Method dispatch page for \code{collect_stats}.
}
\section{Dispatch}{
\code{collect_stats} can be dispatched on following classes:

\itemize{
\item \code{\link{collect_stats,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{collect_stats,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
