\name{colnames-ConsensusPartitionList-method}
\alias{colnames,ConsensusPartitionList-method}
\title{
Column Names of the Matrix
}
\description{
Column Names of the Matrix
}
\usage{
\S4method{colnames}{ConsensusPartitionList}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\examples{
# There is no example
NULL
}
