\name{colnames-dispatch}
\alias{colnames}
\title{
Method dispatch page for colnames
}
\description{
Method dispatch page for \code{colnames}.
}
\section{Dispatch}{
\code{colnames} can be dispatched on following classes:

\itemize{
\item \code{\link{colnames,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
\item \code{\link{colnames,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{colnames,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{colnames,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
