\name{colnames-DownSamplingConsensusPartition-method}
\alias{colnames,DownSamplingConsensusPartition-method}
\title{
Column names of the matrix
}
\description{
Column names of the matrix
}
\usage{
\S4method{colnames}{DownSamplingConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{DownSamplingConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
