\name{compare_signatures-dispatch}
\alias{compare_signatures}
\title{
Method dispatch page for compare_signatures
}
\description{
Method dispatch page for \code{compare_signatures}.
}
\section{Dispatch}{
\code{compare_signatures} can be dispatched on following classes:

\itemize{
\item \code{\link{compare_signatures,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{compare_signatures,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
