\name{config_ATC}
\alias{config_ATC}
\title{
Adjust parameters for default ATC method
}
\description{
Adjust parameters for default ATC method
}
\usage{
config_ATC(cor_fun = stats::cor, min_cor = 0, power = 1, k_neighbours = -1, group = NULL, cores = 1, ...)
}
\arguments{

  \item{cor_fun}{-cor_fun}
  \item{min_cor}{-min_cor}
  \item{power}{-power}
  \item{k_neighbours}{-k_neighbours}
  \item{group}{-group}
  \item{cores}{-cores}
  \item{...}{-...}

}
\examples{
# There is no example
NULL

}
