% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{consensus_heatmap}
\alias{consensus_heatmap}
\title{Heatmap for the consensus matrix}
\usage{
consensus_heatmap(res, k, show_legend = TRUE, annotation = data.frame(known
  = res$known), annotation_color = list(known = res$known_color))
}
\arguments{
\item{res}{a \code{consensus_partition} object.}

\item{k}{number of partitions.}

\item{show_legend}{whether show heatmap and annotation legends.}

\item{annotation}{a data frame with column annotations}

\item{annotation_color}{colors for the annotations}
}
\value{
The consensus matrix.
}
\description{
Heatmap for the consensus matrix
}
