% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspond_rank.R
\name{correspond_between_two_rankings}
\alias{correspond_between_two_rankings}
\title{Correspond two rankings}
\usage{
correspond_between_two_rankings(x1, x2, name1 = "", name2 = "", col1 = 1,
  col2 = 2, top_n = length(x1))
}
\arguments{
\item{x1}{a vector of scores calculated by one metric.}

\item{x2}{a vector of scores calculated by another metric.}

\item{name1}{name of the first metric.}

\item{name2}{name of the second metric.}

\item{col1}{color of the lines for the first metric.}

\item{col2}{color of the lines for the second metric.}

\item{top_n}{top n elements to visualize}
}
\description{
Correspond two rankings
}
\details{
In \code{x1} and \code{x2}, the i^{th} element is the same object but with different
scores under different metrics.
}
\examples{
require(matrixStats)
mat = matrix(runif(1000), ncol = 10)
x1 = rowSds(mat)
x2 = rowMads(mat)
correspond_between_two_rankings(x1, x2, name1 = "sd", name2 = "mad", top_n = 20)
}
