% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{crk_py}
\alias{crk_py}
\title{Create cumulative resistance kernels}
\usage{
crk_py(
  inshp,
  intif,
  outtif,
  maxdist,
  shape,
  transf = "yes",
  volume,
  ncores = as.numeric(Sys.getenv("COLA_NCORES")),
  crs = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/crk.py")
)
}
\arguments{
\item{py}{Python location}

\item{pyscript}{Python script location}
}
\value{
Path with CDPOP results
}
\description{
Cumulative resistant kernels
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
