% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{diagnose_cola}
\alias{diagnose_cola}
\title{Diagnose  \emph{COLA} installation}
\usage{
diagnose_cola(
  envName = "cola",
  libs2Install = c("gdal", "h5py", "numexpr", "rasterio", "pytables", "pandas", "cython",
    "numba", "networkit", "fiona", "shapely", "geopandas", "kdepy", "scikit-image")
)
}
\arguments{
\item{envName}{The environment name. 'cola' by default}

\item{libs2Install}{The name of python libraries}
}
\value{
NULL. Prints in console five (5) logs regarding different steps
}
\description{
Runs some test assessing the correct cola installation
}
\examples{
diagnose_cola()
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
