\name{dim.DownSamplingConsensusPartition}
\alias{dim.DownSamplingConsensusPartition}
\title{
Dimension of the matrix
}
\description{
Dimension of the matrix
}
\usage{
\method{dim}{DownSamplingConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{DownSamplingConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
