\name{dimension_reduction-dispatch}
\alias{dimension_reduction}
\title{
Method dispatch page for dimension_reduction
}
\description{
Method dispatch page for \code{dimension_reduction}.
}
\section{Dispatch}{
\code{dimension_reduction} can be dispatched on following classes:

\itemize{
\item \code{\link{dimension_reduction,matrix-method}}, \code{\link{matrix-class}} class method
\item \code{\link{dimension_reduction,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{dimension_reduction,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
