% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{dimension_reduction}
\alias{dimension_reduction}
\title{Visualize columns after dimension reduction}
\usage{
dimension_reduction(res, k, method = c("mds", "tsne"),
  silhouette_cutoff = 0.5, remove = FALSE, ...)
}
\arguments{
\item{res}{a \code{consensus_partition} object}

\item{k}{number of partitions}

\item{method}{which method to reduce the dimension of the data.}

\item{silhouette_cutoff}{cutoff of silhouette. Data points with values less
than it will be mapped to small points.}

\item{remove}{whether to remove columns which have less silhouette values than
the cutoff.}

\item{...}{pass to \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}}}
}
\description{
Visualize columns after dimension reduction
}
