\name{.onLoad}
\alias{.onLoad}
\alias{Dot.onLoad}
\title{
Global Parameters
}
\description{
Global Parameters
}
\usage{
.onLoad(...)
}
\arguments{

  \item{...}{Arguments for the parameters, see "details" section}
  \item{RESET}{reset to default values}
  \item{READ.ONLY}{please ignore}
  \item{LOCAL}{please ignore}
  \item{ADD}{please ignore}

}
\details{
There are following global parameters:

\describe{
  \item{\code{group_diff}}{Used in \code{\link{get_signatures,ConsensusPartition,method}}.}
}
}
\examples{
# There is no example
NULL
}
