\name{DownSamplingConsensusPartition-class}
\docType{class}
\alias{DownSamplingConsensusPartition}
\alias{DownSamplingConsensusPartition-class}
\title{
The DownSamplingConsensusPartition class
}
\description{
The DownSamplingConsensusPartition class
}
\details{
The \code{DownSamplingConsensusPartition} performs consensus partitioning only with a small subset
of columns and the class of other columns are predicted by \code{\link{predict_classes,ConsensusPartition-method}}.

The \code{DownSamplingConsensusPartition-class} is a child class of \code{\link{ConsensusPartition-class}}. It inherits
all methods of \code{\link{ConsensusPartition-class}}.
}
\seealso{
The constructor function \code{\link{consensus_partition_by_down_sampling}}.
}
\examples{
# There is no example
NULL

}
