\name{enriched_functions_word_cloud}
\alias{enriched_functions_word_cloud}
\title{
Word cloud for the enriched functions
}
\description{
Word cloud for the enriched functions
}
\usage{
enriched_functions_word_cloud(x, genesets, stopwords = GS_STOPWORDS)
}
\arguments{

  \item{x}{the object returned from \code{\link{enrich_signatures_to_genesets}}}
  \item{genesets}{the object also used in \code{\link{enrich_signatures_to_genesets}}}
  \item{stopwords}{words that are not take into account for the word cloud}

}
\examples{
# There is no example
NULL

}
