\name{functional_enrichment-dispatch}
\alias{functional_enrichment}
\title{
Method dispatch page for functional_enrichment
}
\description{
Method dispatch page for \code{functional_enrichment}.
}
\section{Dispatch}{
\code{functional_enrichment} can be dispatched on following classes:

\itemize{
\item \code{\link{functional_enrichment,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{functional_enrichment,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{functional_enrichment,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{functional_enrichment,ANY-method}}, \code{\link{ANY-class}} class method
}
}
\examples{
# no example
NULL


}
