\name{get_anno-ConsensusPartitionList-method}
\alias{get_anno,ConsensusPartitionList-method}
\title{
Get annotations
}
\description{
Get annotations
}
\usage{
\S4method{get_anno}{ConsensusPartitionList}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartitionList-class}} object}

}
\value{
A data frame if \code{anno} was specified in \code{\link{run_all_consensus_partition_methods}}, or \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
