\name{get_anno-dispatch}
\alias{get_anno}
\title{
Method dispatch page for get_anno
}
\description{
Method dispatch page for \code{get_anno}.
}
\section{Dispatch}{
\code{get_anno} can be dispatched on following classes:

\itemize{
\item \code{\link{get_anno,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_anno,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{get_anno,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{get_anno,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
