\name{get_anno-HierarchicalPartition-method}
\alias{get_anno,HierarchicalPartition-method}
\title{
Get annotations
}
\description{
Get annotations
}
\usage{
\S4method{get_anno}{HierarchicalPartition}(object)
}
\arguments{

  \item{object}{a \code{\link{HierarchicalPartition-class}} object}

}
\value{
A data frame if \code{anno} was specified in \code{\link{hierarchical_partition}}, or \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
