\name{get_anno_col-ConsensusPartition-method}
\alias{get_anno_col,ConsensusPartition-method}
\title{
Get annotation colors
}
\description{
Get annotation colors
}
\usage{
\S4method{get_anno_col}{ConsensusPartition}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}

}
\value{
A list of color vectors or \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
