\name{get_anno_col-dispatch}
\alias{get_anno_col}
\title{
Method dispatch page for get_anno_col
}
\description{
Method dispatch page for \code{get_anno_col}.
}
\section{Dispatch}{
\code{get_anno_col} can be dispatched on following classes:

\itemize{
\item \code{\link{get_anno_col,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{get_anno_col,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{get_anno_col,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
