\name{get_anno_col-HierarchicalPartition-method}
\alias{get_anno_col,HierarchicalPartition-method}
\title{
Get annotation colors
}
\description{
Get annotation colors
}
\usage{
\S4method{get_anno_col}{HierarchicalPartition}(object)
}
\arguments{

  \item{object}{A \code{\link{HierarchicalPartition-class}} object.}

}
\value{
A list of color vectors or \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
