\name{get_best_k-ConsensusPartitionList-method}
\alias{get_best_k,ConsensusPartitionList-method}
\title{
Get the best number of partitions
}
\description{
Get the best number of partitions
}
\usage{
\S4method{get_best_k}{ConsensusPartitionList}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartitionList-class}} object}

}
\value{
A data frame with best k for each combination of methods
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
