\name{get_class-dispatch}
\alias{get_class}
\title{
Method dispatch page for get_class
}
\description{
Method dispatch page for \code{get_class}.
}
\section{Dispatch}{
\code{get_class} can be dispatched on following classes:

\itemize{
\item \code{\link{get_class,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_class,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
