\name{get_class-HierarchicalPartition-method}
\alias{get_class,HierarchicalPartition-method}
\title{
Get class from the HierarchicalPartition object
}
\description{
Get class from the HierarchicalPartition object
}
\usage{
\S4method{get_class}{HierarchicalPartition}(object, depth = NULL)
}
\arguments{

  \item{object}{a \code{\link{HierarchicalPartition-class}} object}
  \item{depth}{minimal depth of the hierarchy}

}
\value{
A vector of predicted classes.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
