% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{get_class.consensus_partition}
\alias{get_class.consensus_partition}
\title{Get class from the consensus_partition object}
\usage{
\method{get_class}{consensus_partition}(x, k, ...)
}
\arguments{
\item{x}{a \code{consensus_parittion} object}

\item{k}{number of partitions}

\item{...}{other arguments.}
}
\value{
A data frame with class IDs and other columns.
}
\description{
Get class from the consensus_partition object
}
