% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{get_class.consensus_partition_all_methods}
\alias{get_class.consensus_partition_all_methods}
\title{Get class from the consensus_partition_all_methods object}
\usage{
\method{get_class}{consensus_partition_all_methods}(x, k, ...)
}
\arguments{
\item{x}{a \code{consensus_partition_all_methods} object}

\item{k}{number of partitions}

\item{...}{other arguments}
}
\value{
A data frame with class IDs and other columns.
}
\description{
Get class from the consensus_partition_all_methods object
}
\details{
The class IDs is re-calculated by merging class IDs from all methods.
}
