% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{get_class}
\alias{get_class}
\title{General method for get_class}
\usage{
get_class(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{other arguments}
}
\description{
General method for get_class
}
