% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{get_class.run_all}
\alias{get_class.run_all}
\title{Get class from the run_all object}
\usage{
\method{get_class}{run_all}(x, k, ...)
}
\arguments{
\item{x}{a \code{run_all} object}

\item{k}{number of partitions}

\item{...}{other arguments}
}
\value{
A data frame with class IDs and other columns.
}
\description{
Get class from the run_all object
}
\details{
The class IDs is re-calculated by merging class IDs from all methods.
}
