\name{get_classes-dispatch}
\alias{get_classes}
\title{
Method dispatch page for get_classes
}
\description{
Method dispatch page for \code{get_classes}.
}
\section{Dispatch}{
\code{get_classes} can be dispatched on following classes:

\itemize{
\item \code{\link{get_classes,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{get_classes,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{get_classes,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
