\name{get_hierarchy}
\alias{get_hierarchy}
\title{
Get the partition hierarchy
}
\description{
Get the partition hierarchy
}
\usage{
get_hierarchy(object, depth = max_depth(object))
}
\arguments{

  \item{object}{A \code{\link{HierarchicalPartition-class}} object.}
  \item{depth}{Depth of the hierarchy.}

}
\value{
A \code{\link[stats]{dendrogram}} object.
}
\examples{
# There is no example
NULL
}
