\name{get_matrix-dispatch}
\alias{get_matrix}
\title{
Method dispatch page for get_matrix
}
\description{
Method dispatch page for \code{get_matrix}.
}
\section{Dispatch}{
\code{get_matrix} can be dispatched on following classes:

\itemize{
\item \code{\link{get_matrix,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_matrix,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{get_matrix,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
