\name{get_matrix-DownSamplingConsensusPartition-method}
\alias{get_matrix,DownSamplingConsensusPartition-method}
\title{
Get the original matrix
}
\description{
Get the original matrix
}
\usage{
\S4method{get_matrix}{DownSamplingConsensusPartition}(object, reduce = FALSE)
}
\arguments{

  \item{object}{A \code{\link{DownSamplingConsensusPartition-class}} object.}
  \item{reduce}{Whether to return the reduced matrix where columns are randomly sampled.}

}
\value{
A numeric matrix
}
\examples{
# There is no example
NULL

}
