\name{get_membership-dispatch}
\alias{get_membership}
\title{
Method dispatch page for get_membership
}
\description{
Method dispatch page for \code{get_membership}.
}
\section{Dispatch}{
\code{get_membership} can be dispatched on following classes:

\itemize{
\item \code{\link{get_membership,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_membership,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
