\name{get_signatures-dispatch}
\alias{get_signatures}
\title{
Method dispatch page for get_signatures
}
\description{
Method dispatch page for \code{get_signatures}.
}
\section{Dispatch}{
\code{get_signatures} can be dispatched on following classes:

\itemize{
\item \code{\link{get_signatures,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_signatures,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
\item \code{\link{get_signatures,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
