% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{get_signatures}
\alias{get_signatures}
\title{Get signature rows}
\usage{
get_signatures(res, k, silhouette_cutoff = 0.5, fdr_cutoff = 0.05,
  scale_rows = TRUE, annotation = data.frame(known = res$known),
  annotation_color = list(known = res$known_color), show_legend = TRUE,
  show_column_names = FALSE, ...)
}
\arguments{
\item{res}{A \code{consensus_partition} class object. The object can be returned
from \code{\link[=get_single_run]{get_single_run()}}.}

\item{k}{number of partitions}

\item{silhouette_cutoff}{cutoff for silhouette values. Columns with values
less than it are not used for finding signature rows.}

\item{fdr_cutoff}{cutoff for fdr of the difference between subgroups.}

\item{scale_rows}{whether apply row scaling when making the heatmap.}

\item{annotation}{a data frame which contains annotation of columns.}

\item{annotation_color}{colors for the annotations.}

\item{show_legend}{whether draw the legends on the heatmap.}

\item{show_column_names}{whether show column names on the heatmap.}

\item{...}{other arguments}
}
\value{
A list of three elements:
-\code{mat} the matrix for the signatures
-\code{fdr} FDR for rows
-\code{gropu} subgroups that the rows are significant for
}
\description{
Get signature rows
}
\details{
Basically the function apply test for the difference of subgroups in every
row. Also, to call it a signature for a given subgroup, the values in the
corresponding subgroup should have the highest mean value compared to all
other subgroups. The minimal p-value compared to all other subgroups is taken
as the p-value of the row and used for FDR calculation.
}
