\name{get_single_run-dispatch}
\alias{get_single_run}
\title{
Method dispatch page for get_single_run
}
\description{
Method dispatch page for \code{get_single_run}.
}
\section{Dispatch}{
\code{get_single_run} can be dispatched on following classes:

\itemize{
\item \code{\link{get_single_run,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{get_single_run,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
