\name{get_single_run-HierarchicalPartition-method}
\alias{get_single_run,HierarchicalPartition-method}
\title{
Get result for a specified level in the partition hierarchy
}
\description{
Get result for a specified level in the partition hierarchy
}
\usage{
\S4method{get_single_run}{HierarchicalPartition}(object, node = "0")
}
\arguments{

  \item{object}{a \code{\link{HierarchicalPartition-class}} object}
  \item{node}{node labal, see \code{\link{hierarchical_partition}} for explaination.}

}
\value{
A \code{\link{ConsensusPartition-class}} object.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
