% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{get_single_run}
\alias{get_single_run}
\title{Get object for a single combination of top method and partition method}
\usage{
get_single_run(res_list, top_method = res_list$top_method[1],
  partition_method = res_list$partition_method[1])
}
\arguments{
\item{res_list}{object returned from \code{\link[=run_all]{run_all()}}.}

\item{top_method}{a single string which is used in \code{\link[=run_all]{run_all()}}}

\item{partition_method}{a single string which is used in \code{\link[=run_all]{run_all()}}}
}
\value{
a \code{consensus_partition} class object.
}
\description{
Get object for a single combination of top method and partition method
}
