\name{get_stat-dispatch}
\alias{get_stat}
\title{
Method dispatch page for get_stat
}
\description{
Method dispatch page for \code{get_stat}.
}
\section{Dispatch}{
\code{get_stat} can be dispatched on following classes:

\itemize{
\item \code{\link{get_stat,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{get_stat,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
