\name{get_stats-dispatch}
\alias{get_stats}
\title{
Method dispatch page for get_stats
}
\description{
Method dispatch page for \code{get_stats}.
}
\section{Dispatch}{
\code{get_stats} can be dispatched on following classes:

\itemize{
\item \code{\link{get_stats,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{get_stats,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
