% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{getMnMx}
\alias{getMnMx}
\title{Get MinMax value from raster layer}
\usage{
getMnMx(rastPath, na.rm = TRUE)
}
\arguments{
\item{rastPath}{File location or raster layer. Accepts file path or SpatRaster object}
}
\value{
Numeric vector of two numbers
}
\description{
Extracts the min and max value from raster
}
\examples{
input_tif <- system.file(package = 'cola', 'sampledata/sampleTif.tif')
getMnMx(input_tif)
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
