\name{GO_enrichment-dispatch}
\alias{GO_enrichment}
\title{
Method dispatch page for GO_enrichment
}
\description{
Method dispatch page for \code{GO_enrichment}.
}
\section{Dispatch}{
\code{GO_enrichment} can be dispatched on following classes:

\itemize{
\item \code{\link{GO_enrichment,ANY-method}}, \code{\link{ANY-class}} class method
\item \code{\link{GO_enrichment,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{GO_enrichment,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
