% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{guessNoData}
\alias{guessNoData}
\title{Get NA value from raster layer metadata}
\usage{
guessNoData(path)
}
\arguments{
\item{path}{File location}
}
\value{
String with the NA value. No converted to number to avoid characters
}
\description{
Extracts the NoData value from raster metadata
}
\examples{
input_tif <- system.file(package = 'cola', 'sampledata/sampleTif.tif')
guessNoData(input_tif)
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
