% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_cond_env}
\alias{install_cond_env}
\title{Install \emph{cola} conda environment}
\usage{
install_cond_env(envName, useYML = TRUE, ymlFile = NULL)
}
\arguments{
\item{envName}{The environment name. 'cola' by default}

\item{ymlFile}{Path to YML to use. Default is NULL}
}
\value{
NULL. Prints in console logs regarding different steps
}
\description{
Installs `cola` conda environment
}
\examples{
install_cond_env(envName = 'cola',
    ymlFile = system.file('python/python_conda_config.yml', package = "cola"))
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
