% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{install_conda_env}
\alias{install_conda_env}
\title{Install \emph{cola} conda environment}
\usage{
install_conda_env(
  envName,
  useYML = FALSE,
  ymlFile = NULL,
  packagess = NULL,
  pv = "3.12.11"
)
}
\arguments{
\item{envName}{The environment name. 'cola' by default}

\item{ymlFile}{Path to YML to use. Default is NULL}

\item{packagess}{String. Vector of packagess to be installed. Default is NULL}

\item{pv}{String. Python version to use. Default value is '3.12.11'}
}
\value{
NULL. Prints in console logs regarding different steps
}
\description{
Installs `cola` conda environment
}
\examples{
install_conda_env(envName = 'cola',
    ymlFile = system.file('python/python_conda_config.yml', package = "cola"))
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@nau.edu>
}
