\name{is_best_k-dispatch}
\alias{is_best_k}
\title{
Method dispatch page for is_best_k
}
\description{
Method dispatch page for \code{is_best_k}.
}
\section{Dispatch}{
\code{is_best_k} can be dispatched on following classes:

\itemize{
\item \code{\link{is_best_k,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{is_best_k,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
