\name{is_leaf_node}
\alias{is_leaf_node}
\title{
Test whether a node is a leaf node
}
\description{
Test whether a node is a leaf node
}
\usage{
is_leaf_node(x, node)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}
  \item{node}{A node ID.}

}
\examples{
# There is no example
NULL
}
