\name{is_stable_k-ConsensusPartition-method}
\alias{is_stable_k,ConsensusPartition-method}
\title{
Test whether the current k corresponds to a stable partition
}
\description{
Test whether the current k corresponds to a stable partition
}
\usage{
\S4method{is_stable_k}{ConsensusPartition}(object, k, ...)
}
\arguments{

  \item{object}{A \code{\link{ConsensusPartition-class}} object.}
  \item{k}{Number of subgroups.}
  \item{...}{Pass to \code{\link{suggest_best_k,ConsensusPartition-method}}.}

}
\value{
Logical scalar
}
\examples{
# There is no example
NULL

}
