\name{knitr_add_tab_item}
\alias{knitr_add_tab_item}
\title{
Add one JavaScript tab in the report
}
\description{
Add one JavaScript tab in the report
}
\usage{
knitr_add_tab_item(code, header, desc = "", opt = NULL, message = NULL)
}
\arguments{

  \item{code}{R code to execute}
  \item{header}{header or the title for the tab}
  \item{desc}{decription in the tab}
  \item{opt}{options for knitr chunk}

}
\details{
This function in only for internal use.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
