\name{knitr_insert_tabs}
\alias{knitr_insert_tabs}
\title{
Generate the HTML code for the JavaScript tabs.
}
\description{
Generate the HTML code for the JavaScript tabs.
}
\usage{
knitr_insert_tabs()
}
\details{
This function is only for internal use.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
