% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{lcc_compare_py}
\alias{lcc_compare_py}
\title{Compare maps of least cost paths}
\usage{
lcc_compare_py(
  intif,
  intifs,
  outcsvabs,
  outcsvrel,
  outpngabs,
  outpngrel,
  outfolder,
  inshp = "None",
  shpfield = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/lcc_compare.py"),
  cml = TRUE,
  show.result = TRUE
)
}
\arguments{
\item{intif}{String. File path to the input file used as reference raster.}

\item{intifs}{String. File path to files to be compared. A single string is required
with the complete path and no spaces, as "/c/path/a.csv,/c/path/b.csv,...'}

\item{outcsvabs}{String. File path to the csv output absolute difference table}

\item{outcsvrel}{String. File path to the csv output relative difference table}

\item{outpngabs}{String. File path to the png output absolute difference plot}

\item{outpngrel}{String. File path to the png output relative difference plot}

\item{outfolder}{String. File path to the folder where pairwise comparison between
´intif´ and ´intifs´}

\item{inshp}{String. File path to the shapefile used to mask the comparisson}

\item{shpfield}{String. Field to be used for regionalize the comparisson.}

\item{py}{Python executable location}

\item{pyscript}{Python script location}

\item{cml}{Logical. Print the back-end command line? Default TRUE}

\item{show.result}{Logical. Print the command line result? Default TRUE}
}
\value{
List of two slots: 'file' with the 'outpngabs' folder if success, 
and 'log' with any resulting message.
}
\description{
This tool compares the least cost paths
}
\examples{
#'library(cola)
outdir <- tempdir()
tifs2compare <- paste(sep = ',',
 system.file(package = 'cola', 'sampledata/corridors.tif'),
 system.file(package = 'cola', 'sampledata/corridors_short.tif'),
 system.file(package = 'cola', 'sampledata/corridors_long.tif')
 )
comp_lcc <- lcc_compare_py(
intif = system.file(package = 'cola', 'sampledata/kernels.tif'),
intifs = tifs2compare,
outcsvabs = file.path(outdir, 'abs_table.csv'),
outcsvrel = file.path(outdir, 'rel_table.csv'),
outpngabs = file.path(outdir, 'abs_plot.png'),
outpngrel = file.path(outdir, 'rel_plot.png'),
 outfolder = outdir
)
}
\author{
Patrick Jantz <Patrick.Jantz@gmail.com>

Ivan Gonzalez <ig299@nau.edu>
}
