% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{lccJoblib_py}
\alias{lccJoblib_py}
\title{Create least cost corridors using parallel computing}
\usage{
lccJoblib_py(
  inshp,
  intif,
  outtif,
  maxdist,
  smooth,
  tolerance,
  ncores = as.numeric(Sys.getenv("COLA_NCORES")),
  crs = "None",
  maxram = 6,
  tempFolder = NULL,
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/lcc_joblib.py")
)
}
\arguments{
\item{py}{Python location}

\item{pyscript}{Python script location}
}
\value{
Path with CDPOP results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
