% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{mean_cophcor}
\alias{mean_cophcor}
\title{Cophenetic correlation coefficient in all methods}
\usage{
mean_cophcor(res_list, top_method = res_list$top_method,
  partition_method = res_list$partition_method)
}
\arguments{
\item{res_list}{a \code{run_all} object}

\item{top_method}{a vector of top methods}

\item{partition_method}{a vector of partition methods}
}
\description{
Cophenetic correlation coefficient in all methods
}
