% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{membership_heatmap}
\alias{membership_heatmap}
\title{Heatmap of membership of columns in each random sampling}
\usage{
membership_heatmap(res, k, show_legend = TRUE, annotation = data.frame(known
  = res$known), annotation_color = list(known = res$known_color))
}
\arguments{
\item{res}{a \code{consensus_partition} object}

\item{k}{number of partitions.}

\item{show_legend}{whether show heatmap and annotation legends.}

\item{annotation}{a data frame with column annotations}

\item{annotation_color}{colors for the annotations}
}
\value{
The membership matrix.
}
\description{
Heatmap of membership of columns in each random sampling
}
