\name{merge_node-HierarchicalPartition-method}
\alias{merge_node,HierarchicalPartition-method}
\title{
Merge node
}
\description{
Merge node
}
\usage{
\S4method{merge_node}{HierarchicalPartition}(object, node_id)
}
\arguments{

  \item{object}{-object}
  \item{node_id}{-node_id}

}
\alias{merge_node}
\examples{
# There is no example
NULL
}
