\name{merge_node_param}
\alias{merge_node_param}
\title{
Parameters to merge branches in subgroup dendrogram.
}
\description{
Parameters to merge branches in subgroup dendrogram.
}
\usage{
merge_node_param(depth = Inf, min_n_signatures = -Inf,
    min_p_signatures = -Inf, node_height = -Inf)
}
\arguments{

  \item{depth}{Depth of the dendrogram.}
  \item{min_n_signatures}{Minimal number of signatures for the partitioning on each node.}
  \item{min_p_signatures}{Minimal fraction of sigatures compared to the total number of rows on each node.}
  \item{node_height}{The height of the sub-dendrogram to cut}

}
\examples{
# There is no example
NULL

}
