\name{msigdb_catalogue}
\alias{msigdb_catalogue}
\title{
Get a catelogue from the whole msigdb database
}
\description{
Get a catelogue from the whole msigdb database
}
\usage{
msigdb_catalogue(x, category = "H", sub_category, organism = "Homo sapiens")
}
\arguments{

  \item{x}{the msigdb object from \code{\link{load_msigdb}}}
  \item{category}{category}
  \item{sub_category}{sub category}
  \item{organism}{organism}

}
\value{
a \code{msigdb} object
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
