\name{ncol-ConsensusPartitionList-method}
\alias{ncol,ConsensusPartitionList-method}
\title{
Number of Columns in the Matrix
}
\description{
Number of Columns in the Matrix
}
\usage{
\S4method{ncol}{ConsensusPartitionList}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\examples{
# There is no example
NULL

}
