\name{ncol-dispatch}
\alias{ncol}
\title{
Method dispatch page for ncol
}
\description{
Method dispatch page for \code{ncol}.
}
\section{Dispatch}{
\code{ncol} can be dispatched on following classes:

\itemize{
\item \code{\link{ncol,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{ncol,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
\item \code{\link{ncol,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{ncol,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
