\name{ncol-DownSamplingConsensusPartition-method}
\alias{ncol,DownSamplingConsensusPartition-method}
\title{
Number of columns in the matrix
}
\description{
Number of columns in the matrix
}
\usage{
\S4method{ncol}{DownSamplingConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{DownSamplingConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
