\name{ncol-HierarchicalPartition-method}
\alias{ncol,HierarchicalPartition-method}
\title{
Number of columns in the matrix
}
\description{
Number of columns in the matrix
}
\usage{
\S4method{ncol}{HierarchicalPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
