\name{node_level-HierarchicalPartition-method}
\alias{node_level,HierarchicalPartition-method}
\alias{node_level}
\title{
Information on the nodes
}
\description{
Information on the nodes
}
\usage{
\S4method{node_level}{HierarchicalPartition}(object)
}
\arguments{

  \item{object}{A \code{\link{HierarchicalPartition-class}} object.}

}
\details{
It is the same as \code{\link{node_info,HierarchicalPartition-method}}.
}
\examples{
# There is no example
NULL

}
