\name{nrow-ConsensusPartition-method}
\alias{nrow,ConsensusPartition-method}
\title{
Number of Rows in the Matrix
}
\description{
Number of Rows in the Matrix
}
\usage{
\S4method{nrow}{ConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL
}
