\name{nrow-ConsensusPartitionList-method}
\alias{nrow,ConsensusPartitionList-method}
\title{
Number of rows in the matrix
}
\description{
Number of rows in the matrix
}
\usage{
\S4method{nrow}{ConsensusPartitionList}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\examples{
# There is no example
NULL
}
