\name{nrow-dispatch}
\alias{nrow}
\title{
Method dispatch page for nrow
}
\description{
Method dispatch page for \code{nrow}.
}
\section{Dispatch}{
\code{nrow} can be dispatched on following classes:

\itemize{
\item \code{\link{nrow,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{nrow,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{nrow,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
