\name{nrow-HierarchicalPartition-method}
\alias{nrow,HierarchicalPartition-method}
\title{
Number of rows in the matrix
}
\description{
Number of rows in the matrix
}
\usage{
\S4method{nrow}{HierarchicalPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
