\name{PAC_origin}
\alias{PAC_origin}
\title{
The proportion of ambiguous clustering (PAC score)
}
\description{
The proportion of ambiguous clustering (PAC score)
}
\usage{
PAC_origin(consensus_mat, x1 = 0.1, x2 = 0.9)
}
\arguments{

  \item{consensus_mat}{a consensus matrix.}
  \item{x1}{lower bound to define "ambiguous clustering".}
  \item{x2}{upper bound to define "ambihuous clustering".}

}
\details{
This is the original implementation of PAC method.
}
\examples{
# There is no example
NULL

}
