\name{plot_ecdf-ConsensusPartition-method}
\alias{plot_ecdf,ConsensusPartition-method}
\alias{plot_ecdf}
\title{
Plot the ecdf of the consensus matrix
}
\description{
Plot the ecdf of the consensus matrix
}
\usage{
\S4method{plot_ecdf}{ConsensusPartition}(object, ...)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object.}
  \item{...}{other arguments.}

}
\examples{
# There is no example
NULL

}
