% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Plot the ecdf of the consensus matrix}
\usage{
plot_ecdf(res, ...)
}
\arguments{
\item{res}{a \code{consensus_partition} object.}

\item{...}{other arguments.}
}
\description{
Plot the ecdf of the consensus matrix
}
