% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{points_py}
\alias{points_py}
\title{Transforms suitability to resistance surface}
\usage{
points_py(
  intif,
  outshp,
  param3,
  param4,
  param5,
  param6 = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/create_source_points.py")
)
}
\arguments{
\item{py}{Python location}
}
\value{
Path with CDPOP results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
