\name{predict_classes-dispatch}
\alias{predict_classes}
\title{
Method dispatch page for predict_classes
}
\description{
Method dispatch page for \code{predict_classes}.
}
\section{Dispatch}{
\code{predict_classes} can be dispatched on following classes:

\itemize{
\item \code{\link{predict_classes,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{predict_classes,matrix-method}}, \code{\link{matrix-class}} class method
}
}
\examples{
# no example
NULL


}
