% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{pri_py}
\alias{pri_py}
\title{Runs prioritization}
\usage{
pri_py(
  tif,
  incrk,
  inlcc,
  maskedcsname = paste0(tempfile(), ".tif"),
  outshppoint,
  outshppol,
  outshppatch,
  outtifpatch,
  outtif,
  param7 = 0.5,
  param8 = 1000,
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/prioritize_core_conn.py")
)
}
\arguments{
\item{py}{Python location}
}
\value{
Path with CDPOP results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
