% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{print.consensus_partition}
\alias{print.consensus_partition}
\title{Print the consensus_partition object}
\usage{
\method{print}{consensus_partition}(x, ...)
}
\arguments{
\item{x}{a \code{consensus_partition} object}

\item{...}{other arguments}
}
\description{
Print the consensus_partition object
}
