% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{print.consensus_partition_all_methods}
\alias{print.consensus_partition_all_methods}
\title{Print the consensus_partition_all_methods object}
\usage{
\method{print}{consensus_partition_all_methods}(x, ...)
}
\arguments{
\item{x}{a \code{consensus_partition_all_methods} object}

\item{...}{other arguments}
}
\description{
Print the consensus_partition_all_methods object
}
