\name{print.hc_table_suggest_best_k}
\alias{print.hc_table_suggest_best_k}
\title{
Print the hc_table_suggest_best_k object
}
\description{
Print the hc_table_suggest_best_k object
}
\usage{
\method{print}{hc_table_suggest_best_k}(x, ...)
}
\arguments{

  \item{x}{A \code{hc_table_suggest_best_k} object from \code{\link{suggest_best_k,HierarchicalPartition-method}}.}
  \item{...}{Other arguments.}

}
\examples{
# There is no example
NULL

}
