\name{print.msigdb}
\alias{print.msigdb}
\title{
Print the msigdb class object
}
\description{
Print the msigdb class object
}
\usage{
\method{print}{msigdb}(x, ...)
}
\arguments{

  \item{x}{the \code{msigdb} object}
  \item{...}{other arguments}

}
\value{
No value is returned
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
