% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{print.run_all}
\alias{print.run_all}
\title{Print the run_all object}
\usage{
\method{print}{run_all}(x, ...)
}
\arguments{
\item{x}{a \code{run_all} object}

\item{...}{other arguments}
}
\description{
Print the run_all object
}
