% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{prio_py}
\alias{prio_py}
\title{Runs prioritization}
\usage{
prio_py(
  intif,
  incrk,
  inlcc,
  maskedcsname = paste0(tempfile(), ".tif"),
  outshppoint,
  outshppol,
  outshppatch,
  outtifpatch,
  outtif,
  threshold = 0.5,
  tolerance = 1000,
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/prioritize_core_conn.py"),
  cml = TRUE,
  show.result = TRUE
)
}
\arguments{
\item{maskedcsname}{String. Temporal mask raster file}

\item{outshppoint}{String. Path to the resulting centroid corridors shapefile}

\item{outshppol}{String. Path to the resulting corridors shapefile}

\item{outshppatch}{String. Path to the resulting patches shapefile}

\item{outtifpatch}{String. Path to the resulting patches tif}

\item{outtif}{String. Path to the resulting corridors raster}

\item{threshold}{Decimal. Numeric value between zero and one (0 - 1) to convert continuous kernels into discrete patches.}

\item{tolerance}{Numeric. This is the distance beyond the least-cost path that an animal might traverse when moving between source points. Larger values result in wider corridors.}

\item{py}{Python executable location. Default is `Sys.getenv("COLA_PYTHON_PATH")`}

\item{pyscript}{Python script location. Default is `system.file(package = 'cola', 'python/prioritize_core_conn.py')`}

\item{cml}{Logical. Print the back-end command line? Default TRUE}

\item{show.result}{Logical. Print the command line result? Default TRUE}
}
\value{
Path with CDPOP results
}
\description{
Run corridors prioritization algorithm. It takes the kernels and corridors as arguments.
}
\examples{
library(cola)
outdir <- tempdir()
prioritization <- prio_py(intif = system.file(package = 'cola', 'sampledata/sampleSR.tif'),
                          incrk = system.file(package = 'cola', 'sampledata/kernels.tif'),
                          inlcc = system.file(package = 'cola', 'sampledata/corridors.tif'),
                          threshold = 0.7,  tolerance = 100000,
                          maskedcsname = file.path(outdir, 'pri_mask.tif'),
                          outshppoint = file.path(outdir, 'pri_points.shp'),
                          outshppol = file.path(outdir, 'pri_polygon.shp'),
                          outshppatch = file.path(outdir, 'pri_patch.shp'),
                          outtifpatch = file.path(outdir, 'pri_patcht.tif'),
                          outtif = file.path(outdir, 'corridors_zarr.tif'))

}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
