% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{quotepath}
\alias{quotepath}
\title{Quote file path if space is detected}
\usage{
quotepath(path)
}
\arguments{
\item{path}{The file path to be quoted}
}
\description{
This function add double quotes to the string if a space is detected
}
\examples{
library(cola)
quotepath(path = c('C:/Users/First Second Name/Documents', 'C:/Users/FirstSecondName/Documents'))
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
