\name{recalc_stats}
\alias{recalc_stats}
\title{
Recalculate statistics in the ConsensusPartitionList object
}
\description{
Recalculate statistics in the ConsensusPartitionList object
}
\usage{
recalc_stats(rl)
}
\arguments{

  \item{rl}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\details{
It updates the statistics slot in the ConsensusPartitionList object, used internally.
}
\examples{
# There is no example
NULL

}
