% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{remove_partition_method}
\alias{remove_partition_method}
\title{Remove partition methods}
\usage{
remove_partition_method(method)
}
\arguments{
\item{method}{name of the partition methods to be removed.}
}
\description{
Remove partition methods
}
