\name{remove_top_value_method}
\alias{remove_top_value_method}
\title{
Remove top methods
}
\description{
Remove top methods
}
\usage{
remove_top_value_method(method)
}
\arguments{

  \item{method}{name of the top methods to be removed.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
