\name{remove_top_value_methods}
\alias{remove_top_value_methods}
\title{
Remove top-value methods
}
\description{
Remove top-value methods
}
\usage{
remove_top_value_methods(method)
}
\arguments{

  \item{method}{Name of the top-value methods to be removed.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
