\name{rownames-ConsensusPartition-method}
\alias{rownames,ConsensusPartition-method}
\title{
Row Names of the Matrix
}
\description{
Row Names of the Matrix
}
\usage{
\S4method{rownames}{ConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
