\name{rownames-dispatch}
\alias{rownames}
\title{
Method dispatch page for rownames
}
\description{
Method dispatch page for \code{rownames}.
}
\section{Dispatch}{
\code{rownames} can be dispatched on following classes:

\itemize{
\item \code{\link{rownames,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{rownames,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
