% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run subgroup classification in a batch}
\usage{
run_all(data, top_method = ALL_TOP_VALUE_METHOD(),
  partition_method = ALL_PARTITION_METHOD(), mc.cores = 1,
  get_signatures = TRUE, ...)
}
\arguments{
\item{data}{a numeric matrix where subgroups are found by columns.}

\item{top_method}{method which are used to extract top n rows. Allowed methods
are in \code{\link[=ALL_TOP_VALUE_METHOD]{ALL_TOP_VALUE_METHOD()}} and can be self-added by \code{\link[=register_top_value_fun]{register_top_value_fun()}}.}

\item{partition_method}{method which are used to do partition on data columns.
Allowed methods are in \code{\link[=ALL_PARTITION_METHOD]{ALL_PARTITION_METHOD()}} and can be self-added
by \code{\link[=register_partition_fun]{register_partition_fun()}}.}

\item{mc.cores}{number of cores to use.}

\item{get_signatures}{whether to run \code{\link[=get_signatures]{get_signatures()}} for each partition.}

\item{...}{other arguments passed to \code{\link[=consensus_partition]{consensus_partition()}}.}
}
\value{
a \code{consensus_partition_all_methods} class object. Following methods can be used on it: \code{\link[=collect_plots]{collect_plots()}},
\code{\link[=collect_classes]{collect_classes()}}, \code{\link[=get_class]{get_class()}}, \code{\link[=get_single_run]{get_single_run()}}.
}
\description{
Run subgroup classification in a batch
}
