% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{select_k}
\alias{select_k}
\title{Several plots for determine the optimized number of partitions}
\usage{
select_k(res, plot = 0)
}
\arguments{
\item{res}{a \code{consensus_partition} object}

\item{plot}{\itemize{
\item 0: plot all four plots;
- 1: only the ecdf plot;
- 2: plot cophenetic correlation coefficient;
- 3: mean silhouette;
- 4: proportion increase of the AUC of the ecdf.
}}
}
\description{
Several plots for determine the optimized number of partitions
}
