% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_cola}
\alias{setup_cola}
\title{Install  \emph{COLA}}
\usage{
setup_cola(
  envName = "cola",
  nSteps = 5,
  force = FALSE,
  yml = TRUE,
  libs2Install = c("gdal", "h5py", "numexpr", "rasterio", "pytables", "pandas", "cython",
    "numba", "networkit", "fiona", "shapely", "geopandas", "kdepy", "scikit-image",
    "kdepy")
)
}
\arguments{
\item{envName}{The environment name. 'cola' by default}

\item{nSteps}{The number of steps for printing log in console}

\item{force}{Force miniconda installation? Passed to `reticulate::install_miniconda()`}

\item{yml}{Use YML file to build the conda environment? Default TRUE}

\item{libs2Install}{The name of python libraries}
}
\value{
NULL. Prints in console logs regarding different steps
}
\description{
Installs cola software. Includes `reticulate` R package, miniconda, cola conda environment and libraries,
and defines the ways R and python should interact
}
\examples{
setup_cola()
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
