% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cola2.R
\name{setup_cola2}
\alias{setup_cola2}
\title{Install  \emph{COLA}}
\usage{
setup_cola2(
  envName = "cola",
  nSteps = 5,
  force = FALSE,
  yml = FALSE,
  onlyIndividual = TRUE,
  ask = TRUE,
  dss = FALSE,
  libs2Install = c("geopandas", "gdal", "h5py", "numexpr", "rasterio", "pytables",
    "pandas", "cython", "numba", "networkit==11.0", "fiona", "shapely", "kdepy",
    "scikit-image", "kdepy"),
  cola2 = FALSE,
  libs2Install2 = c("google-api-python-client", "google-auth", "google-cloud-storage",
    "geemap", "numpy", "matplotlib", "scikit-learn", "requests", "pandas", "geopandas",
    "shapely")
)
}
\arguments{
\item{envName}{The environment name. 'cola' by default}

\item{nSteps}{The number of steps for printing log in console}

\item{force}{Force miniconda installation? Passed to `reticulate::install_miniconda()`}

\item{yml}{Use YML file to build the conda environment? Default TRUE}

\item{onlyIndividual}{Try installing libraries one by one? Default FALSE}

\item{ask}{Ask before installing reticulate, minoconda, and cola conda environment?. Default TRUE}

\item{dss}{Install cola DSS as well? Default FALSE}

\item{libs2Install}{The name of python libraries}
}
\value{
NULL. Prints in console logs regarding different steps
}
\description{
Installs cola software. Includes `reticulate` R package, miniconda, cola conda environment and libraries,
and defines the ways R and python should interact
}
\examples{
setup_cola()
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
