\name{show-ConsensusPartition-method}
\alias{show,ConsensusPartition-method}
\title{
Print the consensus_partition object
}
\description{
Print the consensus_partition object
}
\usage{
\S4method{show}{ConsensusPartition}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}
  \item{...}{other arguments}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
