\name{show-dispatch}
\alias{show}
\title{
Method dispatch page for show
}
\description{
Method dispatch page for \code{show}.
}
\section{Dispatch}{
\code{show} can be dispatched on following classes:

\itemize{
\item \code{\link{show,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{show,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{show,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{show,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
}
}
\examples{
# no example
NULL


}
