\name{show-DownSamplingConsensusPartition-method}
\alias{show,DownSamplingConsensusPartition-method}
\title{
Print the DownSamplingConsensusPartition object
}
\description{
Print the DownSamplingConsensusPartition object
}
\usage{
\S4method{show}{DownSamplingConsensusPartition}(object)
}
\arguments{

  \item{object}{A \code{\link{DownSamplingConsensusPartition-class}} object.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
data(golub_cola_ds)
golub_cola_ds
}
