\name{signature_density-ConsensusPartition-method}
\alias{signature_density,ConsensusPartition-method}
\title{
Density for the signatures
}
\description{
Density for the signatures
}
\usage{
\S4method{signature_density}{ConsensusPartition}(object, k, ...)
}
\arguments{

  \item{object}{A \code{\link{ConsensusPartition-class}} object. }
  \item{k}{number of partitions}
  \item{...}{pass to \code{\link{get_signatures,ConsensusPartition-method}}}

}
\details{
The function makes density distributio nf of signatures in all columns.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\alias{signature_density}
\examples{
# There is no example
NULL
}
