\name{suggest_best_k-dispatch}
\alias{suggest_best_k}
\title{
Method dispatch page for suggest_best_k
}
\description{
Method dispatch page for \code{suggest_best_k}.
}
\section{Dispatch}{
\code{suggest_best_k} can be dispatched on following classes:

\itemize{
\item \code{\link{suggest_best_k,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{suggest_best_k,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{suggest_best_k,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
