\name{test_to_known_factors-dispatch}
\alias{test_to_known_factors}
\title{
Method dispatch page for test_to_known_factors
}
\description{
Method dispatch page for \code{test_to_known_factors}.
}
\section{Dispatch}{
\code{test_to_known_factors} can be dispatched on following classes:

\itemize{
\item \code{\link{test_to_known_factors,ConsensusPartition-method}}, \code{\link{ConsensusPartition-class}} class method
\item \code{\link{test_to_known_factors,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
\item \code{\link{test_to_known_factors,DownSamplingConsensusPartition-method}}, \code{\link{DownSamplingConsensusPartition-class}} class method
\item \code{\link{test_to_known_factors,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
}
}
\examples{
# no example
NULL


}
