\name{top_rows_heatmap-dispatch}
\alias{top_rows_heatmap}
\title{
Method dispatch page for top_rows_heatmap
}
\description{
Method dispatch page for \code{top_rows_heatmap}.
}
\section{Dispatch}{
\code{top_rows_heatmap} can be dispatched on following classes:

\itemize{
\item \code{\link{top_rows_heatmap,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{top_rows_heatmap,matrix-method}}, \code{\link{matrix-class}} class method
}
}
\examples{
# no example
NULL


}
