% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{top_rows_heatmap}
\alias{top_rows_heatmap}
\title{Heatmap for the top rows}
\usage{
top_rows_heatmap(res_list, top_n = 2000)
}
\arguments{
\item{res_list}{a \code{run_all} object}

\item{top_n}{number of top rows}
}
\description{
Heatmap for the top rows
}
