\name{top_rows_overlap-dispatch}
\alias{top_rows_overlap}
\title{
Method dispatch page for top_rows_overlap
}
\description{
Method dispatch page for \code{top_rows_overlap}.
}
\section{Dispatch}{
\code{top_rows_overlap} can be dispatched on following classes:

\itemize{
\item \code{\link{top_rows_overlap,ConsensusPartitionList-method}}, \code{\link{ConsensusPartitionList-class}} class method
\item \code{\link{top_rows_overlap,matrix-method}}, \code{\link{matrix-class}} class method
\item \code{\link{top_rows_overlap,HierarchicalPartition-method}}, \code{\link{HierarchicalPartition-class}} class method
}
}
\examples{
# no example
NULL


}
