% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{top_rows_overlap}
\alias{top_rows_overlap}
\title{Overlap of top rows from different top methods}
\usage{
top_rows_overlap(res_list, top_n = 2000, type = c("venn", "correspondance"))
}
\arguments{
\item{res_list}{a \code{consensus_partition_all_methods} object}

\item{top_n}{number of top rows}

\item{type}{venn: use venn euler plots; correspondance: use \code{\link[=correspond_between_rankings]{correspond_between_rankings()}}.}
}
\description{
Overlap of top rows from different top methods
}
