% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_partition.R
\name{venn_euler}
\alias{venn_euler}
\title{Make Venn Euler diagram from a list}
\usage{
venn_euler(lt, ...)
}
\arguments{
\item{lt}{a list of items}

\item{...}{other arguments}
}
\description{
Make Venn Euler diagram from a list
}
