\name{all_leaves-HierarchicalPartition-method}
\alias{all_leaves,HierarchicalPartition-method}
\alias{all_leaves}
\title{
all leaves in the hierarchy
}
\description{
all leaves in the hierarchy
}
\usage{
\S4method{all_leaves}{HierarchicalPartition}(object, depth = NULL)
}
\arguments{

  \item{object}{the object}
  \item{depth}{depth}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
