\name{all_nodes-HierarchicalPartition-method}
\alias{all_nodes,HierarchicalPartition-method}
\title{
all nodes in the hierarchy
}
\description{
all nodes in the hierarchy
}
\usage{
\S4method{all_nodes}{HierarchicalPartition}(object, depth = NULL)
}
\arguments{

  \item{object}{the object}
  \item{depth}{depth}

}
\alias{all_nodes}
\examples{
# There is no example
NULL
}
