\name{all_partition_methods}
\alias{all_partition_methods}
\title{
All supported partition methods
}
\description{
All supported partition methods
}
\usage{
all_partition_methods()
}
\value{
A vector of supported partition methods
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
all_partition_methods()
}
