% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{cdmat_py}
\alias{cdmat_py}
\title{Creates CDmatrix for CDPOP model}
\usage{
cdmat_py(
  inshp,
  intif,
  outcsv,
  param4,
  param5 = Sys.getenv("COLA_NCORES"),
  param6 = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/create_cdmat.py")
)
}
\arguments{
\item{py}{Python location}
}
\value{
Path with CDPOP results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
