% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{cdpop_py}
\alias{cdpop_py}
\title{Run CDPOP model}
\usage{
cdpop_py(
  py = Sys.getenv("COLA_PYTHON_PATH"),
  cdpopscript = system.file(package = "cola", "CDPOP/src/CDPOP.py"),
  inputvars = NULL,
  agevars = NULL,
  cdmat = NULL,
  xy = NULL,
  tempFolder,
  prefix = paste0("cdpopout", sessionIDgen(only3 = TRUE))
)
}
\arguments{
\item{py}{String. Python location}

\item{cdpopscript}{String. Python location}

\item{inputvars}{String.}

\item{agevars}{String.}

\item{cdmat}{String.}

\item{xy}{String.}

\item{prefix}{String.}

\item{tempfolder}{String.}
}
\value{
List with two slots: a) newFiles with generated results, b) cdpopPath, with the folder with the results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
