% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_params.R
\docType{data}
\name{cola_params}
\alias{cola_params}
\title{Parameters for installation}
\format{
An object of class \code{list} of length 5.
}
\usage{
cola_params
}
\value{
The temperature in degrees Celsius
}
\description{
Default parameters for cola conda environment installation
}
\keyword{datasets}
