\name{cola_report-HierarchicalPartition-method}
\alias{cola_report,HierarchicalPartition-method}
\title{
Make report for the HierarchicalPartition object
}
\description{
Make report for the HierarchicalPartition object
}
\usage{
\S4method{cola_report}{HierarchicalPartition}(object, output_dir, env = parent.frame())
}
\arguments{

  \item{object}{a \code{\link{HierarchicalPartition-class}} object}
  \item{output_dir}{the path for the output html file}
  \item{env}{where object is found, internally used}

}
\details{
A html report which contains all plots
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
