\name{colnames-ConsensusPartition-method}
\alias{colnames,ConsensusPartition-method}
\title{
Column Names of the Matrix
}
\description{
Column Names of the Matrix
}
\usage{
\S4method{colnames}{ConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL
}
