\name{colnames-HierarchicalPartition-method}
\alias{colnames,HierarchicalPartition-method}
\title{
Column names of the matrix
}
\description{
Column names of the matrix
}
\usage{
\S4method{colnames}{HierarchicalPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}

}
\examples{
# There is no example
NULL
}
