% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{crk_compare_py}
\alias{crk_compare_py}
\title{Compare maps of cumulative resistance kernels}
\usage{
crk_compare_py(
  intif,
  intifs,
  outcsvabs,
  outcsvrel,
  outpngabs,
  outpngrel,
  outfolder,
  inshp = "None",
  shpfield = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/crk_compare.py")
)
}
\arguments{
\item{py}{Python location}
}
\value{
Path with CDPOP results
}
\description{
This tool compares the cumulative resistance kernels
}
\examples{
crk_compare_py( )
}
\author{
Patrick Jantz <Patrick.Jantz@gmail.com>

Ivan Gonzalez <ig299@nau.edu>
}
