\name{dim.ConsensusPartitionList}
\alias{dim.ConsensusPartitionList}
\title{
Dimension of the Matrix
}
\description{
Dimension of the Matrix
}
\usage{
\method{dim}{ConsensusPartitionList}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\examples{
# There is no example
NULL

}
