\name{dim.HierarchicalPartition}
\alias{dim.HierarchicalPartition}
\title{
Dimension of the Matrix
}
\description{
Dimension of the Matrix
}
\usage{
\method{dim}{HierarchicalPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}

}
\examples{
# There is no example
NULL
}
