\name{[.ConsensusPartitionList}
\alias{[.ConsensusPartitionList}
\title{
Subset a ConsensusPartitionList object
}
\description{
Subset a ConsensusPartitionList object
}
\usage{
\method{[}{ConsensusPartitionList}(x, i, j, drop = TRUE)
}
\arguments{

  \item{x}{a \code{\link{ConsensusPartition-class}} object}
  \item{i}{index for top value methods}
  \item{j}{index for partition methods}
  \item{drop}{whether drop class}

}
\details{
A
}
\examples{
# There is no example
NULL

}
