\name{[[.HierarchicalPartition}
\alias{[[.HierarchicalPartition}
\title{
Subset
}
\description{
Subset
}
\usage{
\method{[[}{HierarchicalPartition}(x, i)
}
\arguments{

  \item{x}{a \code{\link{HierarchicalPartition-class}} object}
  \item{i}{index}

}
\examples{
data(cola_rh)
cola_rh[["01"]]
cola_rh[["2"]]
}
