\name{get_anno_col-ConsensusPartitionList-method}
\alias{get_anno_col,ConsensusPartitionList-method}
\title{
Get annotation colors
}
\description{
Get annotation colors
}
\usage{
\S4method{get_anno_col}{ConsensusPartitionList}(object)
}
\arguments{

  \item{object}{A \code{\link{ConsensusPartitionList-class}} object}

}
\value{
A list of color vectors or else \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
