\name{get_consensus-ConsensusPartition-method}
\alias{get_consensus,ConsensusPartition-method}
\alias{get_consensus}
\title{
Get consensus matrix
}
\description{
Get consensus matrix
}
\usage{
\S4method{get_consensus}{ConsensusPartition}(object, k)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}
  \item{k}{number of partitions}

}
\examples{
# There is no example
NULL

}
