\name{get_membership-ConsensusPartition-method}
\alias{get_membership,ConsensusPartition-method}
\alias{get_membership}
\title{
Get membership matrix
}
\description{
Get membership matrix
}
\usage{
\S4method{get_membership}{ConsensusPartition}(object, k, each = FALSE)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}
  \item{k}{number of partitions}
  \item{each}{return the percentage membership matrix or the membership in every random round}

}
\examples{
# There is no example
NULL

}
