\name{get_stat-ConsensusPartition-method}
\alias{get_stat,ConsensusPartition-method}
\alias{get_stat}
\title{
Get parameters
}
\description{
Get parameters
}
\usage{
\S4method{get_stat}{ConsensusPartition}(object, k = object@k)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}
  \item{k}{number of partitions. The value can be a vector.}

}
\examples{
# There is no example
NULL

}
