\name{get_stat-ConsensusPartitionList-method}
\alias{get_stat,ConsensusPartitionList-method}
\title{
Get parameters
}
\description{
Get parameters
}
\usage{
\S4method{get_stat}{ConsensusPartitionList}(object, k)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartitionList-class}} object}
  \item{k}{number of partitions.}

}
\examples{
# There is no example
NULL

}
