\name{knitr_add_tab_item}
\alias{knitr_add_tab_item}
\title{
Add one tab item
}
\description{
Add one tab item
}
\usage{
knitr_add_tab_item(code, header, desc = "", opt = NULL)
}
\arguments{

  \item{code}{R code}
  \item{header}{header for the tab}
  \item{desc}{decription}
  \item{opt}{options for knitr}

}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
