% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{lcc_py}
\alias{lcc_py}
\title{Corridors}
\usage{
lcc_py(
  inshp,
  intif,
  outtif,
  maxdist,
  smooth,
  tolerance,
  ncores = as.numeric(Sys.getenv("COLA_NCORES")),
  crs = "None",
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/lcc.py")
)
}
\arguments{
\item{py}{Python location}

\item{pyscript}{Python script location}
}
\value{
Path with CDPOP results
}
\description{
Calculate corridors
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
