% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{lccHeavy_py}
\alias{lccHeavy_py}
\title{Create least cost corridors for heavy rasters}
\usage{
lccHeavy_py(
  inshp,
  intif,
  outtif,
  param4,
  param5,
  param6,
  param7 = as.numeric(Sys.getenv("COLA_NCORES")),
  param8 = "None",
  tempFolder = rootPath,
  py = Sys.getenv("COLA_PYTHON_PATH"),
  pyscript = system.file(package = "cola", "python/lcc_heavy.py")
)
}
\arguments{
\item{py}{Python location}
}
\value{
Path with CDPOP results
}
\description{
Run CDPOP model
}
\examples{
runCDPOP( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
