\name{load_msigdb}
\alias{load_msigdb}
\title{
Load from MSigDB
}
\description{
Load from MSigDB
}
\usage{
load_msigdb(f)
}
\arguments{

  \item{f}{path of the xml file.}

}
\details{
The xml file can be downloaded from \url{http://software.broadinstitute.org/gsea/downloads.jsp} .
}
\examples{
# There is no example
NULL

}
