\name{max_depth-HierarchicalPartition-method}
\alias{max_depth,HierarchicalPartition-method}
\alias{max_depth}
\title{
max depth of the hierarchy
}
\description{
max depth of the hierarchy
}
\usage{
\S4method{max_depth}{HierarchicalPartition}(object)
}
\arguments{

  \item{object}{the object}

}
\examples{
# There is no example
NULL

}
