\name{merge_node-HierarchicalPartition-method}
\alias{merge_node,HierarchicalPartition-method}
\alias{merge_node}
\title{
Merge node
}
\description{
Merge node
}
\usage{
\S4method{merge_node}{HierarchicalPartition}(object, node_id)
}
\arguments{

  \item{object}{A \code{\link{HierarchicalPartition-class}} object.}
  \item{node_id}{A vector of node IDs where each node is merged as a leaf node.}

}
\value{
A \code{\link{HierarchicalPartition-class}} object.
}
\examples{
# There is no example
NULL

}
