\name{ncol-ConsensusPartition-method}
\alias{ncol,ConsensusPartition-method}
\title{
Number of Columns in the Matrix
}
\description{
Number of Columns in the Matrix
}
\usage{
\S4method{ncol}{ConsensusPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
