\name{PAC}
\alias{PAC}
\title{
PAC scores
}
\description{
PAC scores
}
\usage{
PAC(consensus_mat)
}
\arguments{

  \item{consensus_mat}{consensus matrix}

}
\details{
This a variant of the orignial PAC (proportion of ambiguous clustering) method with 
slight modification.
}
\value{
A single PAC score
}
\examples{
# There is no example
NULL

}
