\name{plot_ecdf-ConsensusPartition-method}
\alias{plot_ecdf,ConsensusPartition-method}
\alias{plot_ecdf}
\title{
Plot the ecdf of the consensus matrix
}
\description{
Plot the ecdf of the consensus matrix
}
\usage{
\S4method{plot_ecdf}{ConsensusPartition}(object, lwd = 4, ...)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object.}
  \item{lwd}{line width}
  \item{...}{other arguments.}

}
\details{
This function is mainly used in \code{\link{collect_plots}} function.
}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
