% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{randPtsFun}
\alias{randPtsFun}
\title{Create random points}
\usage{
randPtsFun(rvect, npts, rmin, rmax)
}
\arguments{
\item{rvect}{Python location}

\item{npts}{Python location}

\item{rmin}{Python location}

\item{rmax}{Python location}
}
\value{
Path with CDPOP results
}
\description{
Create random points
}
\examples{
hs <- system.file(package = 'cola', 'sampledata/sampleTif.tif')
pts30 <- s2res_py(intif = hs, outtif = 'hs_p3_0.tif', param3 = 0, param4 = 1, param5 = 10,  param6 = 1,  param7 = NULL, param8 = 'None')
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
