\name{register_NMF}
\alias{register_NMF}
\title{
Register NMF partition method
}
\description{
Register NMF partition method
}
\usage{
register_NMF(package = c("NMF", "NNLM"))
}
\arguments{

  \item{package}{which package to support NMF.}

}
\details{
Note \code{\link[NMF]{nmf}} is very time-consuming.
}
\examples{
# There is no example
NULL

}
