\name{register_NMF}
\alias{register_NMF}
\title{
Register NMF partition method
}
\description{
Register NMF partition method
}
\usage{
register_NMF(package = c("NMF", "NNLM"))
}
\arguments{

  \item{package}{which package to support NMF, \code{NMF} package or \code{NNLM} package.}

}
\details{
Note NMF analysis is very time-consuming.
}
\examples{
# There is no example
NULL

}
