\name{remove_partition_method}
\alias{remove_partition_method}
\title{
Remove partition methods
}
\description{
Remove partition methods
}
\usage{
remove_partition_method(method)
}
\arguments{

  \item{method}{name of the partition methods to be removed.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL
}
