\name{remove_partition_method}
\alias{remove_partition_method}
\title{
Remove partition methods
}
\description{
Remove partition methods
}
\usage{
remove_partition_method(method)
}
\arguments{

  \item{method}{name of the partition methods to be removed.}

}
\examples{
# There is no example
NULL

}
