% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{remove_top_value_method}
\alias{remove_top_value_method}
\title{Remove top methods}
\usage{
remove_top_value_method(method)
}
\arguments{
\item{method}{name of the top methods to be removed.}
}
\description{
Remove top methods
}
