% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{replacePixels}
\alias{replacePixels}
\title{Add values to  maps of least cost paths}
\usage{
replacePixels(
  polPath,
  burnval = "val2burn",
  rastPath,
  colu = FALSE,
  att = FALSE,
  rastCRS = NA,
  gdal = TRUE
)
}
\arguments{
\item{burnval}{String. Value to burn. Can be a number or a attribute/column name}

\item{rastPath}{String. Location of the raster layer}

\item{att}{Logical. Should be 'all-the-touched' pixels be considered? Default TRUE}

\item{polpath}{String. Location of the vector layer}

\item{lineBuffW}{Number. How many pixels should be used as buffer width for line geometries?}
}
\value{
Path of the resulting raster layer. Same as rastPath with the '_replaced ' suffix.
}
\description{
Rasterize a polygon and sum it to an existing raster. Both layers need to be in the same projection.
}
\examples{
replacePixels( )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
