\name{rownames-ConsensusPartitionList-method}
\alias{rownames,ConsensusPartitionList-method}
\title{
Row names of the matrix
}
\description{
Row names of the matrix
}
\usage{
\S4method{rownames}{ConsensusPartitionList}(x)
}
\arguments{

  \item{x}{A \code{\link{ConsensusPartitionList-class}} object.}

}
\examples{
# There is no example
NULL
}
