\name{rownames-HierarchicalPartition-method}
\alias{rownames,HierarchicalPartition-method}
\title{
Row Names of the Matrix
}
\description{
Row Names of the Matrix
}
\usage{
\S4method{rownames}{HierarchicalPartition}(x)
}
\arguments{

  \item{x}{A \code{\link{HierarchicalPartition-class}} object.}

}
\examples{
# There is no example
NULL

}
