\name{select_partition_number-ConsensusPartition-method}
\alias{select_partition_number,ConsensusPartition-method}
\alias{select_partition_number}
\title{
Several plots for determine the optimized number of partitions
}
\description{
Several plots for determine the optimized number of partitions
}
\usage{
\S4method{select_partition_number}{ConsensusPartition}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartition-class}} object}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
