\name{show-ConsensusPartition-method}
\alias{show,ConsensusPartition-method}
\title{
Print the ConsensusPartition object
}
\description{
Print the ConsensusPartition object
}
\usage{
\S4method{show}{ConsensusPartition}(object)
}
\arguments{

  \item{object}{A \code{\link{ConsensusPartition-class}} object.}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
