\name{show-ConsensusPartitionList-method}
\alias{show,ConsensusPartitionList-method}
\title{
Print the ConsensusPartitionList object
}
\description{
Print the ConsensusPartitionList object
}
\usage{
\S4method{show}{ConsensusPartitionList}(object)
}
\arguments{

  \item{object}{a \code{\link{ConsensusPartitionList-class}} object}
  \item{...}{other arguments}

}
\examples{
# There is no example
NULL

}
