\name{show-HierarchicalPartition-method}
\alias{show,HierarchicalPartition-method}
\title{
Print the HierarchicalPartition object
}
\description{
Print the HierarchicalPartition object
}
\usage{
\S4method{show}{HierarchicalPartition}(object)
}
\arguments{

  \item{object}{A \code{\link{HierarchicalPartition-class}} object}

}
\value{
No value is returned.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
data(cola_rh)
cola_rh
}
