% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{shp2xy}
\alias{shp2xy}
\title{Shapefile to CDPOP xy}
\usage{
shp2xy(shapefile, outxy, tempDir)
}
\arguments{
\item{shapefile}{String. Path to the file}
}
\description{
Convert Shapefile into xy file
}
\examples{
shp2xy( shapefile = 'shapefilepathhere.shp',outxy = 'out.xy', tempDir = 'temFolder')
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
