% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cola_fun.R
\name{shp2xy}
\alias{shp2xy}
\title{shapefile to CDPOP XY}
\usage{
shp2xy(
  shapefile,
  outxy,
  tempDir,
  mortrast = NULL,
  survrast = NULL,
  porcEmpty = 0
)
}
\arguments{
\item{shapefile}{String. Path to the shapefile}

\item{outxy}{String. Path to the resulting .xy file}

\item{tempDir}{String. Path to the working directory}

\item{mortrast}{String. Optional. Path to mortality/resistance raster. If it is a valid  raster, extracts the values to populate the 'Subpop_mortperc' field. The extracted values are scaled between 0 and 100.}

\item{survrast}{String. Optional. Path to survival/suitability raster. This argument will be ignored if a valid mortrast is provided. If it is a valid  raster, extracts the values to populate the 'Subpop_mortperc' field. The extracted values are flipped upside down and scaled between 0 and 100.}

\item{porcEmpty}{Integer. percentage of locations (points) that randomly will be considered empty and possible to be colonized.}
}
\value{
String. Path to a temporal xy.csv file. Writes the same file at the outxy path
}
\description{
Converts Shapefile into xy file
}
\examples{
(tempfolder <- tempfile())
dir.create(tempfolder)

(xyfile <- shp2xy( shapefile = system.file(package = 'cola', 'sampledata/points_sabah_50.shp'),
        outxy = file.path(out_dir, 'out.xy'), tempDir = tempfolder))
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
