% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cola_functions.R
\name{test_cola}
\alias{test_cola}
\title{Test cola functions}
\usage{
test_cola(run = FALSE, zarr = FALSE)
}
\arguments{
\item{run}{Logical. Run the tests?}

\item{zarr}{Logical. Run zarr corridors? Default FALSE}
}
\value{
Path with CDPOP results
}
\description{
Run the main geospatial CoLa tools: Suitability to Resistance, Kernels, Corridors
}
\examples{
library(cola)
test_cola(run = TRUE )
}
\author{
Ivan Gonzalez <ig299@nau.edu>

Patrick Jantz <Patrick.Jantz@gmail.com>
}
